requirements/base.txt
amqp==2.3.2
billiard==3.5.0.4
celery==4.2.1
Django==2.1.2
kombu==4.2.1
Pillow==5.3.0
pytz==2018.5
vine==1.1.4
sorl-thumbnail==12.5.0
djangorestframework==3.9.0
django-paypal==0.5.0
pycryptodome==3.7.2
social-auth-app-django==3.1.0